@@ -6,27 +6,27 @@
     "require": {
         "php": "^8.1",
         "ext-simplexml": "*",
-        "doctrine/collections": "^1.5 || ^2.0",
+        "doctrine/collections": "^2.0",
         "eonx-com/easy-core": "^4.4",
         "eonx-com/easy-utils": "^4.4",
-        "nesbot/carbon": "^2.22",
-        "nette/utils": "^3.1",
-        "phpunit/phpunit": "^9.5 || ^10.0",
-        "symfony/config": "^5.4 || ^6.0",
-        "symfony/console": "^5.4 || ^6.0",
-        "symfony/contracts": "^2.0 || ^3.0",
-        "symfony/dependency-injection": "^5.4 || ^6.0",
-        "symfony/event-dispatcher": "^5.4 || ^6.0",
-        "symfony/framework-bundle": "^5.4 || ^6.0",
-        "symfony/http-client": "^5.4 || ^6.0",
-        "symfony/http-kernel": "^5.4 || ^6.0",
-        "symfony/intl": "^5.4 || ^6.0",
-        "symfony/mailer": "^5.4 || ^6.0",
-        "symfony/messenger": "^5.4 || ^6.0",
-        "symfony/property-access": "^5.4 || ^6.0",
-        "symfony/translation": "^5.4 || ^6.0",
-        "symfony/validator": "^5.4 || ^6.0",
-        "symfony/yaml": "^5.4 || ^6.0"
+        "nesbot/carbon": "^2.67",
+        "nette/utils": "^3.2",
+        "phpunit/phpunit": "^9.6 || ^10.2",
+        "symfony/config": "^6.3",
+        "symfony/console": "^6.3",
+        "symfony/contracts": "^3.3",
+        "symfony/dependency-injection": "^6.3",
+        "symfony/event-dispatcher": "^6.3",
+        "symfony/framework-bundle": "^6.3",
+        "symfony/http-client": "^6.3",
+        "symfony/http-kernel": "^6.3",
+        "symfony/intl": "^6.3",
+        "symfony/mailer": "^6.3",
+        "symfony/messenger": "^6.3",
+        "symfony/property-access": "^6.3",
+        "symfony/translation": "^6.3",
+        "symfony/validator": "^6.3",
+        "symfony/yaml": "^6.3"
     },
     "autoload": {
         "psr-4": {
