@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace EonX\EasyTest\Console;
 
-use EonX\EasyUtils\CollectorHelper;
+use EonX\EasyUtils\Helpers\CollectorHelper;
 use Symfony\Component\Console\Application;
 use Symfony\Component\Console\Command\Command;
 
